+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Fix(es):
 
 - Phase out old Benchmark records by marking as INACTIVE.
+- Refactor `threads` config field to a `multithread` bool in `compression_text` benchmark scores.
 
 ## v0.4.0 (Mar 14, 2026)