+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 Fix(es):
 
 - Phase out old Benchmark records by marking as INACTIVE.
-- Use `lsblk` storage data for `GCP` and `OVH` servers, with `lshw` as a fallback.
+- Use `lsblk` storage data for `GCP` servers, with `lshw` as a fallback.
+- Fix `OVH` storage info parsing from API data.
 
 ## v0.4.0 (Mar 14, 2026)